consolidate common code for role/attribute filters
